+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - FIX kumo Parameter diffing does not work for aws coma-seperated inputs (#77).
 - FIX kumo Parameter diffing does not work for aws coma-seperated inputs (#77).
 - FEATURE tenkai now has the slack notifications (#79).- FIX moved tests to pytest to improve cleanup after tests (#119).
+- FIX moved tests to pytest to improve cleanup after tests (#119).
+- FIX ramuda rollback to previous version.
 
 ## [0.0.57] - 2016-09-23
 ### Added
@@ -74,3 +76,5 @@ to bundle packages at runtime.
 - FIX: removed debug output
 - FIX: removed some unnecessary import validations
 - FIX: kumo will now exit when importing a cloudformation.py not from your current working directory
+
+...